// src/main/resources/angular/src/app/api/model/privacyPolicy.ts
/**
 * MAAS API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PrivacyPolicyType } from './privacyPolicyType';


export interface PrivacyPolicy {
    privacyPoliciesId: number;
    version: number;
    active: number;
    publishDate: string;
    /**
     * Url with the locale of the user. To be used by the apps.
     */
    url?: string;
    type: PrivacyPolicyType;
    url_ca: string;
    url_es: string;
    url_en: string;
    client: string;
}